token-extended/server/config.py
import os
from dotenv import load_dotenv
from datetime import timedelta

basedir = os.path.abspath(os.path.dirname(__file__))

load_dotenv()

class Config:
    SQLALCHEMY_DATABASE_URI = os.environ.get('DATABASE_URI') \
                              or 'sqlite:///' + os.path.join(basedir, 'app.db')
    SQLALCHEMY_TRACK_MODIFICATIONS = False
    JWT_SECRET_KEY = os.environ.get('JWT_SECRET_KEY') or 'set this key in .env file'
    JWT_TOKEN_LOCATION = ['cookies']
    JWT_COOKIE_SECURE = os.environ.get('JWT_COOKIE_SECURE') or False  # Set to True if using HTTPS only in production
    JWT_COOKIE_CSRF_PROTECT = True  # Enable CSRF protection, default is True - just being explicit
    JWT_ACCESS_TOKEN_EXPIRES = timedelta(minutes=2)
